requirements.txt
flask==2.0.1
Werkzeug==2.0.1
tensorflow-cpu==2.13.1
pillow==9.5.0
numpy==1.24.3
